+++ src/config.js
@@ -5,6 +5,8 @@ export default {
     CARDS_API: process.env.CARDS_API || 'https://data.riskmap.us/cards/',
     CARDS_API_KEY: process.env.CARDS_API_KEY,
     CARDS_URL: process.env.CARDS_URL || 'https://cards.riskmap.us/flood/',
+    DEFAULT_INSTANCE_REGION_CODE: process.env.DEFAULT_INSTANCE_REGION_CODE ||
+    'brw',
     DEFAULT_LANGUAGE: process.env.DEFAULT_LANGUAGE || 'en',
     MAP_URL: process.env.MAP_URL || 'https://riskmap.us/',
     TWITTER_CONSUMER_KEY: process.env.TWITTER_CONSUMER_KEY,