+++ CHANGELOG.md
@@ -16,3 +16,5 @@ v0.0.2
 v0.0.3
 - Refactored to use cognicity-bot-core module
 - Removed serverless requirement
+- Add request signature verification
+- Change to use Twitter buttons